refactor(Engine): cast screen size values to floats for precision
- Updated `Line.cpp` to cast `WindowSize.X` and `WindowSize.Y` to `int` before assignment.
- Updated `Translate` method in `Line.h` to use `[[nodiscard]]` for better code clarity.
- Refactored `main.cpp` to use a more efficient method for generating circle vertices and drawing them using triangles.

+++ Main/Engine/Actors/Line/Line.cpp
@@ -47,14 +47,14 @@ LDDrift::Actors::Line& LDDrift::Actors::Line::SetThickness(const float thick) {
 }
 
 LDDrift::Actors::Line& LDDrift::Actors::Line::SetScreenSize(const LDDrift::VecPos2D& WindowSize) {
-    WidthScreen = WindowSize.X;
-    HeightScreen = WindowSize.Y;
+    WidthScreen = static_cast<int>(WindowSize.X);
+    HeightScreen = static_cast<int>(WindowSize.Y);
     return *this;
 }
 
 std::vector<LDDrift::Actors::Line::PNT> LDDrift::Actors::Line::Translate() const {
-    const float HalfWidth = WidthScreen / 2.0f;
-    const float HalfHeight = HeightScreen / 2.0f;
+    const float HalfWidth = static_cast<float>(WidthScreen) / 2.0f;
+    const float HalfHeight = static_cast<float>(HeightScreen) / 2.0f;
     std::vector<LDDrift::Actors::Line::PNT> CopyPoints = Points;
     CopyPoints[0].Point_1.X = (CopyPoints[0].Point_1.X - HalfWidth) / HalfWidth;
     CopyPoints[0].Point_1.Y = (CopyPoints[0].Point_1.Y - HalfHeight) / HalfHeight;
